+++ src/com/github/tnerevival/commands/admin/AdminPurgeCommand.java
@@ -1,10 +1,13 @@
 package com.github.tnerevival.commands.admin;
 
-import org.bukkit.ChatColor;
-import org.bukkit.command.CommandSender;
-
 import com.github.tnerevival.TNE;
 import com.github.tnerevival.commands.TNECommand;
+import com.github.tnerevival.core.Message;
+import com.github.tnerevival.utils.MISCUtils;
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 
 
 public class AdminPurgeCommand extends TNECommand {
@@ -35,13 +38,18 @@ public class AdminPurgeCommand extends TNECommand {
   
   @Override
   public boolean execute(CommandSender sender, String command, String[] arguments) {
-    boolean world = arguments.length >= 1;
-    //TODO: Messages
-    if(world) {
+    String world = (sender instanceof Player)? MISCUtils.getWorld((Player)sender) : TNE.instance.defaultWorld;
+    boolean isWorld = arguments.length >= 1;
+    if(isWorld) {
+      if(Bukkit.getWorld(arguments[0]) == null)
+        new Message("Messages.General.NoWorld").translate(world, sender);
       TNE.instance.manager.purge(arguments[0]);
+      Message m = new Message("Messages.Admin.PurgeWorld");
+      m.addVariable("$world", arguments[0]);
       return true;
     }
     TNE.instance.manager.purgeAll();
+    new Message("Messages.Admin.Purge").translate(world, sender);
     return true;
   }